nexus/translators/to_python.py
"""
Agent YAML → Python Class Translator

Generates Python class scaffolding from agent.yaml definitions.
Produces a template that developers fill in with actual logic.

Usage:
    python -m nexus.cli translate to-python triage-agent.yaml > triage_agent.py
"""


import yaml


def translate_agent_to_python(yaml_path: str) -> str:
    """Load YAML agent definition and generate Python class."""

    with open(yaml_path) as f:
        agent = yaml.safe_load(f)

    metadata = agent.get("metadata", {})
    spec = agent.get("spec", {})

    class_name = metadata.get("name", "Agent").title().replace("-", "")

    # Generate input type hints
    input_hints = []
    for param_name, param_def in spec.get("inputs", {}).items():
        param_type = param_def.get("type", "str")
        if param_type == "string" or param_type == "enum":
            py_type = "str"
        elif param_type == "array":
            py_type = "list[str]"
        else:
            py_type = param_type
        input_hints.append(f"    {param_name}: {py_type}")

    # Generate output type hints
    output_hints = []
    for output_name, output_def in spec.get("outputs", {}).items():
        output_type = output_def.get("type", "str")
        if output_type == "string" or output_type == "enum":
            py_type = "str"
        elif output_type == "array":
            py_type = "list[str]"
        else:
            py_type = output_type
        output_hints.append(f"    {output_name}: {py_type}")

    py = f'''"""
{metadata.get('name')} - {metadata.get('description')}

Version: {metadata.get('version')}
Author: {metadata.get('author')}

This is a generated class template. Fill in the run() method with your logic.
"""

from dataclasses import dataclass
from typing import Optional
import asyncio


@dataclass
class {class_name}Input:
    """Input contract for {class_name}."""
{chr(10).join(input_hints)}


@dataclass
class {class_name}Output:
    """Output contract for {class_name}."""
{chr(10).join(output_hints)}


class {class_name}:
    """
    {metadata.get('description')}

    Purpose:
    {spec.get('purpose', '')}
    """

    def __init__(self, ai_provider, git_platform):
        """Initialize agent with required dependencies."""
        self.ai = ai_provider
        self.git = git_platform

    async def run(self, inputs: {class_name}Input) -> {class_name}Output:
        """
        Execute the agent.

        This is the main entry point. Implement your agent logic here.

        Available tools:
{chr(10).join('        - {tool}' for tool in spec.get('requires_tools', []))}

        The following prompt is provided as guidance for the LLM:

        {{{{AGENT_INSTRUCTIONS}}}}

        {spec.get('ai_instructions', '').replace('{', '{{').replace('}', '}}')}

        {{{{/AGENT_INSTRUCTIONS}}}}
        """

        # TODO: Implement agent logic here
        # 1. Process inputs
        # 2. Call AI or external tools
        # 3. Construct and return output

        raise NotImplementedError("Implement the run() method in {class_name}")

    async def _call_ai(self, prompt: str) -> str:
        """Helper: Call LLM with formatted prompt."""
        # TODO: Construct proper prompt with input values
        result = await self.ai.complete(prompt)
        return result

    async def _parse_response(self, response: str) -> dict:
        """Helper: Parse LLM response into structured output."""
        # TODO: Parse and validate response format
        import json
        return json.loads(response)


# Example usage
if __name__ == "__main__":
    async def demo():
        # Create agents with your AI and Git providers
        agent = {class_name}(ai_provider=..., git_platform=...)

        # Prepare inputs
        inputs = {class_name}Input(
{chr(10).join('            # TODO: fill in\n' for _ in spec.get('inputs', {}).items())}
        )

        # Run agent
        output = await agent.run(inputs)
        print(output)

    # asyncio.run(demo())
'''

    return py
